backend/requirements.txt
fastapi
uvicorn
easyocr
python-multipart
pillow
numpy
opencv-python
pydantic-settings
